requirements2.txt
Django==5.2.5
mysqlclient==2.2.0
django-bootstrap5==25.2
django-crispy-forms==1.14.0
django-colorfield==0.8.0
django-extensions==3.1.5
django-jazzmin==2.6.0
django-paypal==2.0
djangorestframework==3.14.0
Pillow==9.5.0
stripe==9.8.0
python-slugify==8.0.1
crispy-forms-materialize==0.2
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
pytest==8.2.2
pytest-django==4.8.0
django-allauth
chardet
PyJWT
channels
channels-redis
daphne
meilisearch==0.31.5
django-environ
djangorestframework-simplejwt
pyinstaller
pandas
openpyxl
xhtml2pdf
pycairo
gunicorn
requests
matplotlib
playwright
quickchart.io
